move inventory modules to separate folder
