+++ src/components/inventory-search.ts
@@ -1,7 +1,7 @@
-import { InventoryAction } from '../core/inventory-actions'
-import inventoryActions from '../core/inventory-actions'
 import { Plugin } from '../components/plugin-container'
 import { RowNormal } from '../components/row-container'
+import { InventoryAction } from '../inventory/actions'
+import inventoryActions from '../inventory/actions'
 import { registerShortcut } from '../core/input'
 import Input from '../components/text-input'
 import { VimMode } from '../neovim/types'
