+++ src/components/inventory.ts
@@ -1,8 +1,8 @@
 import { InputMode, switchInputMode, watchInputMode, registerShortcut } from '../core/input'
-import { InventoryAction } from '../core/inventory-actions'
-import { InventoryLayer } from '../core/inventory-layers'
-import inventoryActions from '../core/inventory-actions'
-import inventoryLayers from '../core/inventory-layers'
+import { InventoryAction } from '../inventory/actions'
+import { InventoryLayer } from '../inventory/layers'
+import inventoryActions from '../inventory/actions'
+import inventoryLayers from '../inventory/layers'
 import { VimMode } from '../neovim/types'
 import { h, app } from '../ui/uikit'
 import nvim from '../core/neovim'
