+++ src/inventory/layers.ts
